requirements.txt
django
django-crispy-forms
django-bootstrap4
django-bootstrap-datepicker-plus
requests